feat : 채널 스켈레톤
